requirements.txt
opencage>=1.2.1
pandas>=1.1.1
python-dotenv>=0.14.0